refactor: restructure product feature into packages
