+++ src/main/kotlin/ecommerce/exception/product/DuplicateProductNameException.kt
@@ -1,2 +1,3 @@
-package ecommerce.exception.product 
+package ecommerce.exception.product
 
+class DuplicateProductNameException(message: String) : RuntimeException(message)
