+++ src/main/kotlin/ecommerce/exception/product/ProductNotFoundException.kt
@@ -1,2 +1,3 @@
-package ecommerce.exception.product 
+package ecommerce.exception.product
 
+class ProductNotFoundException(message: String) : RuntimeException(message)
